Fix Windows CI fail due to GH13866 and patch up tests
