+++ .gitignore
@@ -2,6 +2,7 @@
 #  if you want to ignore files created by your editor/tools,
 #  please consider a global .gitignore https://help.github.com/articles/ignoring-files
 *.exe
+*.exe~
 *.orig
 *.rej
 *.test
