+++ integration-cli/docker_cli_port_test.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"fmt"
-	"net"
 	"os/exec"
 	"regexp"
 	"sort"
@@ -146,82 +145,6 @@ func assertPortList(c *check.C, out string, expected []string) bool {
 	}
 
 	return true
-}
-
-func (s *DockerSuite) TestPortHostBinding(c *check.C) {
-	runCmd := exec.Command(dockerBinary, "run", "-d", "-p", "9876:80", "busybox",
-		"nc", "-l", "-p", "80")
-	out, _, err := runCommandWithOutput(runCmd)
-	if err != nil {
-		c.Fatal(out, err)
-	}
-	firstID := strings.TrimSpace(out)
-
-	runCmd = exec.Command(dockerBinary, "port", firstID, "80")
-	out, _, err = runCommandWithOutput(runCmd)
-	if err != nil {
-		c.Fatal(out, err)
-	}
-
-	if !assertPortList(c, out, []string{"0.0.0.0:9876"}) {
-		c.Error("Port list is not correct")
-	}
-
-	runCmd = exec.Command(dockerBinary, "run", "--net=host", "busybox",
-		"nc", "localhost", "9876")
-	if out, _, err = runCommandWithOutput(runCmd); err != nil {
-		c.Fatal(out, err)
-	}
-
-	runCmd = exec.Command(dockerBinary, "rm", "-f", firstID)
-	if out, _, err = runCommandWithOutput(runCmd); err != nil {
-		c.Fatal(out, err)
-	}
-
-	runCmd = exec.Command(dockerBinary, "run", "--net=host", "busybox",
-		"nc", "localhost", "9876")
-	if out, _, err = runCommandWithOutput(runCmd); err == nil {
-		c.Error("Port is still bound after the Container is removed")
-	}
-}
-
-func (s *DockerSuite) TestPortExposeHostBinding(c *check.C) {
-	runCmd := exec.Command(dockerBinary, "run", "-d", "-P", "--expose", "80", "busybox",
-		"nc", "-l", "-p", "80")
-	out, _, err := runCommandWithOutput(runCmd)
-	if err != nil {
-		c.Fatal(out, err)
-	}
-	firstID := strings.TrimSpace(out)
-
-	runCmd = exec.Command(dockerBinary, "port", firstID, "80")
-	out, _, err = runCommandWithOutput(runCmd)
-	if err != nil {
-		c.Fatal(out, err)
-	}
-
-	_, exposedPort, err := net.SplitHostPort(out)
-
-	if err != nil {
-		c.Fatal(out, err)
-	}
-
-	runCmd = exec.Command(dockerBinary, "run", "--net=host", "busybox",
-		"nc", "localhost", strings.TrimSpace(exposedPort))
-	if out, _, err = runCommandWithOutput(runCmd); err != nil {
-		c.Fatal(out, err)
-	}
-
-	runCmd = exec.Command(dockerBinary, "rm", "-f", firstID)
-	if out, _, err = runCommandWithOutput(runCmd); err != nil {
-		c.Fatal(out, err)
-	}
-
-	runCmd = exec.Command(dockerBinary, "run", "--net=host", "busybox",
-		"nc", "localhost", strings.TrimSpace(exposedPort))
-	if out, _, err = runCommandWithOutput(runCmd); err == nil {
-		c.Error("Port is still bound after the Container is removed")
-	}
 }
 
 func stopRemoveContainer(id string, c *check.C) {
