+++ integration-cli/requirements_unix.go
@@ -36,4 +36,22 @@ var (
 		},
 		"Test requires an environment that supports cgroup cfs quota.",
 	}
+	OomControl = TestRequirement{
+		func() bool {
+			cgroupMemoryMountpoint, err := cgroups.FindCgroupMountpoint("memory")
+			if err != nil {
+				return false
+			}
+			if _, err := ioutil.ReadFile(path.Join(cgroupMemoryMountpoint, "memory.memsw.limit_in_bytes")); err != nil {
+				return false
+			}
+
+			if _, err = ioutil.ReadFile(path.Join(cgroupMemoryMountpoint, "memory.oom_control")); err != nil {
+				return false
+			}
+			return true
+
+		},
+		"Test requires Oom control enabled.",
+	}
 )
